examples/plugins/normalizer/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = 'nomad-normalizer-plugin-example'
version = '1.0'
description = 'An example normalizer plugin for NOMAD.'
readme = "README.md"
authors = [{ name = "The NOMAD Authors" }]
license = { text = "Apache-2.0" }
dependencies = [
    "nomad-lab>=1.2.0"
]

[project.urls]
homepage = "https://github.com/nomad-coe/nomad-normalizer-example-plugin"

[project.optional-dependencies]
tests = [
    'pytest'
]

[tool.setuptools.packages.find]
include = ["nomadnormalizerexample*"]

[tool.setuptools.package-data]
nomadnormalizerexample = ['*/nomad_plugin.yaml']
